refactor(calibration): split CalibrationRefitService into runner/queue/job (max-params 2/3)
CalibrationRefitService injected 8 deps (surreal, apiKeys, calibration,
config, jobs, claim, workerLoop, guard). Four-way split:
- CalibrationRefitRunnerService (surreal, calibration, apiKeys) — the
  refit engine (source-trust + isotonic calibration), with per-tenant
  progress callbacks; extractorModel via process.env.
- CalibrationRefitQueueService (apiKeys, claim, workerLoop) — handler
  registration + single cross-tenant cron enqueue + queue-mode flag.
- CalibrationRefitJobService (apiKeys, jobs, guard) — wraps an inline run
  with the distributed lease + job_run lifecycle (start/progress/finish).
- CalibrationRefitService (runner, queue, job) — cron cadence +
  queue-vs-inline routing; public refit*/listVersions delegate.

Pure helpers (aggregateBySourceKey/isCorrect) + persistCalibrationMap
moved to the runner; the bootstrap-key and pure-helper specs retarget it.
Each class ≤3 deps.

Part of the max-params=3 god-class split program.
